requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
humanize>=4.8.0
groq>=0.37.0
PyNaCl>=1.5.0
Pillow>=10.0.0
gTTS>=2.4.0
aiohttp>=3.9.0
static-ffmpeg>=2.5.0
ffmpeg-python>=0.2.0
discord-ext-voice-recv>=0.5.1a170
edge-tts>=6.1.0
elevenlabs>=1.0.0
aioredis>=2.0.0
redis>=5.0.0
audioop-lts>=0.2.1; python_version >= "3.13"
